refactor headers into same translation unit
