--- core/src/main/scala/zio/http/model/AcceptCharset.scala
@@ -1,28 +1,0 @@
-/*
- *
- *  Copyright 2017-2019 John A. De Goes and the ZIO Contributors
- *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
- *
- */
-
-package zio.http.model
-
-/*
-    Accept-Charset header field used to indicate what charsets are
-    acceptable in textual response content.
-
-    E.g. Accept-Charset: iso-8859-5, unicode-1-1;q=0.8
- */
-
-final case class AcceptCharset(charsets: List[Charset])
