--- core/src/main/scala/zio/http/model/AcceptLanguage.scala
@@ -1,30 +1,0 @@
-/*
- *
- *  Copyright 2017-2019 John A. De Goes and the ZIO Contributors
- *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
- *
- */
-
-package zio.http.model
-
-/*
-
-    Accept-Language header field to indicate the set of natural
-    languages that are preferred in the response.
-
-    E.g. Accept-Language: da, en-gb;q=0.8, en;q=0.7
-
- */
-
-final case class AcceptLanguage(languages: List[Language])
