--- core/src/main/scala/zio/http/model/Authorization.scala
@@ -1,8 +1,0 @@
-package zio.http.model
-
-final case class Authorization(
-  scheme: AuthenticationScheme,
-  credentials: Credentials
-) {
-  override def toString = s"Authorization: $scheme $credentials"
-}
