+++ core/src/main/scala/zio/http/model/CharsetRange.scala
@@ -1,0 +1,3 @@
+package zio.http.model
+
+final case class CharsetRange(cs: Charset, q: Option[QValue])
