+++ core/src/main/scala/zio/http/model/Email.scala
@@ -1,0 +1,3 @@
+package zio.http.model
+
+final case class Email(address: String)
